Added check that a folder's files aren't null before attempting to iterate over them

+++ android/app/src/main/java/com/microsoft/codepush/react/FileUtils.java
@@ -72,11 +72,13 @@ public class FileUtils {
     public static void deleteFileOrFolderSilently(File file) {
         if (file.isDirectory()) {
             File[] files = file.listFiles();
-            for (File fileEntry : files) {
-                if (fileEntry.isDirectory()) {
-                    deleteFileOrFolderSilently(fileEntry);
-                } else {
-                    fileEntry.delete();
+            if (files != null) {
+                for (File fileEntry : files) {
+                    if (fileEntry.isDirectory()) {
+                        deleteFileOrFolderSilently(fileEntry);
+                    } else {
+                        fileEntry.delete();
+                    }
                 }
             }
         }
